fix(tracker): getParam returns int32 so magic/range damage records > 65k are accurate
uint16 return type silently truncated any action param above 65,535 (e.g.
200,000 → 3,392). The C++ actionResult_t stores param as int32; the Lua
getter now returns the full value. Fixes MaxNuke/MaxBurst/MaxDmg30 in
combat_records.lua recording the wrong value for hits above 65k.
WS tracking was unaffected (uses weaponskillHit local var, not getParam).
Requires C++ rebuild.

+++ src/map/lua/lua_action.cpp
@@ -70,7 +70,7 @@ void CLuaAction::actionID(const uint16 actionid) const
     m_PLuaAction->actionid = actionid;
 }
 
-auto CLuaAction::getParam(const uint32 actionTargetId) const -> uint16
+auto CLuaAction::getParam(const uint32 actionTargetId) const -> int32
 {
     for (auto&& actionTarget : m_PLuaAction->targets)
     {
